% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANALYSE__functions.R
\name{analyse.behavioural.data}
\alias{analyse.behavioural.data}
\title{Analyse behavioural data}
\usage{
analyse.behavioural.data(bd_df, aggregation_column_list)
}
\arguments{
\item{bd_df}{Behavioural data frame/table}

\item{aggregation_column_list}{List of columns to group by}
}
\value{
Provides behavioural information for the experiment as a data.table.
}
\description{
Analyse behavioural data
}
\examples{
# BREAK UP BY TARGET-PRESENT AND TARGET-ABSENT TRIALS - THE COLUMN TRIALTYPE_TEXT
data(fixationreport)
behaviouralData <- analyse.behavioural.data(fixationreport, aggregation_column_list = list('TRIALTYPE_TEXT'))
}
