% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{AOIName}
\alias{AOIName}
\title{AOIName}
\format{A data frame with 72 rows and 3 variables.
\describe{
\item{section}{name of AOI}
\item{x}{x coordinates of the borders}
\item{y}{y coordinates of the borders}
}}
\usage{
data(AOIName)
}
\description{
The defined Area of Interest information with function \code{DrawAOIs}
}
\keyword{datasets}
