% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{crosshair}
\alias{crosshair}
\title{crosshair}
\format{A data frame with 5 rows and 2 variables.
\describe{
\item{V1}{the x coordinate of the crosshair in the scene.}
\item{V2}{the y coordinate of the crosshair in the scene.}
}}
\usage{
data(crosshair)
}
\description{
The x and y coordinates of the crosshair in the scene.
}
\keyword{datasets}
