% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertImages_DrawAOIs.R
\name{DrawAOIs}
\alias{DrawAOIs}
\title{DrawAOIs}
\usage{
DrawAOIs(boxes, poster, out = "RegionName.csv")
}
\arguments{
\item{boxes}{Define how many area of interests one is going to create.}

\item{poster}{The location and name of the poster image file.}

\item{out}{The name of the output csv file. Default is "RegionName.csv"}
}
\value{
RegionBorder The border information of the defined area of interests.
}
\description{
Definie rectangular Area of Interests by clicking the vertex of the rectangle and type the names in the R console
}
\examples{
\dontrun{
AOIdata <- DrawAOIs(boxes = 2,
poster = system.file("extdata", "poster_colored.jpg", package = "EyeTrackR"))
}
}
\author{
Chunyang Li < lichunyang1990@hotmail.com >
}
