% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherPlotsFunctions.R
\name{DrawEyeAOITimelines}
\alias{DrawEyeAOITimelines}
\title{DrawEyeAOITimelines}
\usage{
DrawEyeAOITimelines(scanpathdat, xlab = "AOIs", ylab = "Time (second)",
  main.title = "Poster AOI Timelines", label.cex = 1.05)
}
\arguments{
\item{scanpathdat}{Summarized scan path eye tracking data with function get.scanpathdat.}

\item{xlab}{Label for x axis. Default is "AOIs".}

\item{ylab}{Label for y axis. Default is "Time (second)".}

\item{main.title}{The title. Default is "Poster AOI Timelines".}

\item{label.cex}{The size of the labels on the x axis. Default is 1.05.}
}
\value{
Returns DrawEyeAOITimelines plot.
}
\description{
Creates an AOI timeline plot indicating the length and the sequence of how the participant looks at each AOIs.
}
\examples{
data(scanpathdat.all)
scanpathdat = scanpathdat.all[[1]]
DrawEyeAOITimelines(scanpathdat = scanpathdat)

}
\author{
Chunyang Li < lichunyang1990@hotmail.com >
}
