% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherPlotsFunctions.R
\name{DrawEyeBarplot}
\alias{DrawEyeBarplot}
\title{DrawEyeBarplot}
\usage{
DrawEyeBarplot(posterdat = posterdat, decreasing = TRUE, xlab = "",
  ylab = "Length of Visits (sec)", ...)
}
\arguments{
\item{posterdat}{Summarized eye tracking data with function GetPosterData.}

\item{decreasing}{TRUE for sorting the data in descending order. Default is TRUE.}

\item{xlab}{Label for x axis. Default is empty.}

\item{ylab}{Label for y axis. Default is 'Length of Visits (sec)'.}

\item{...}{Arguments to be passed to methods, such as graphical parameters in \code{\link[graphics]{barplot}}.}
}
\value{
Returns a bar plot.
}
\description{
Creates a bar plot of how long one is looking at each AOIs.
}
\examples{
data(posterdat.all)
posterdat = posterdat.all[[1]]
DrawEyeBarplot(posterdat = posterdat)
}
\references{
See \url{https://stat.ethz.ch/R-manual/R-devel/library/graphics/html/barplot.html}.
}
\seealso{
\code{\link{GetPosterData}}
}
\author{
Chunyang Li < lichunyang1990@hotmail.com >
}
