% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherPlotsFunctions.R
\name{DrawEyeBoxplot}
\alias{DrawEyeBoxplot}
\title{DrawEyeBoxplot}
\usage{
DrawEyeBoxplot(pupildat = pupildat, decreasing = TRUE, ...)
}
\arguments{
\item{pupildat}{Summarized pupil data with function GetPosterData.}

\item{decreasing}{TRUE for sorting the data in descending order. Default is TRUE.}

\item{...}{Arguments to be passed to methods, such as graphical parameters in \code{\link[graphics]{boxplot}}.}
}
\value{
Returns a box plot of participant's pupil radiuses while looking at each AOIs.
}
\description{
Creates a box plot of participant's pupil radiuses while looking at each AOIs.
}
\examples{
\dontrun{
data(posterdat.all)
pupildat = posterdat.all[[3]]
DrawEyeBoxplot(pupildat = pupildat)
}
}
\references{
See \url{https://stat.ethz.ch/R-manual/R-devel/library/graphics/html/boxplot.html}.
}
\author{
Chunyang Li < lichunyang1990@hotmail.com >
}
