% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherPlotsFunctions.R
\name{DrawEyeDotplot}
\alias{DrawEyeDotplot}
\title{DrawEyeDotplot}
\usage{
DrawEyeDotplot(posterdat = posterdat, decreasing = TRUE,
  xlab = "Area of Interest", ylab = "Length of Visits (sec)", ...)
}
\arguments{
\item{posterdat}{the directory of the video frames}

\item{decreasing}{TRUE for sorting the data in descending order. Default is TRUE.}

\item{xlab}{Label for x axis. Default is 'Area of Interest'.}

\item{ylab}{Label for y axis. Default is 'Length of Visits (sec)'.}

\item{...}{Arguments to be passed to methods, such as graphical parameters in \code{\link[graphics]{plot}}.}
}
\value{
Returns a dot plot.
}
\description{
Extract the coordinates of the crosshair in terms of the poster from the video frames
}
\examples{
data(posterdat.all)
posterdat = posterdat.all[[1]]
DrawEyeDotplot(posterdat = posterdat)

}
\references{
See \url{https://stat.ethz.ch/R-manual/R-devel/library/graphics/html/plot.html}.
}
\seealso{
\code{\link{GetPosterData}}
}
\author{
Chunyang Li < lichunyang1990@hotmail.com >
}
