% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherPlotsFunctions.R
\name{DrawEyeHeatmap}
\alias{DrawEyeHeatmap}
\title{DrawEyeHeatmap}
\usage{
DrawEyeHeatmap(poster, locations, bandwidth = c(60, 60), main = "",
  adjust = TRUE)
}
\arguments{
\item{poster}{The location and name of the poster image file.}

\item{locations}{The directory and name of the csv file of the eye tracking data coordinates.}

\item{bandwidth}{The bandwidth of the heatmap. The default value is c(60, 60).}

\item{main}{The title. Default is empty.}

\item{adjust}{Adjust the difference in the coordinates from the results returned by Matlab. Default is TRUE.}
}
\value{
Returns a heatmap.
}
\description{
Creates a heat map with Gaussian kernel function showing where the participant focus most.
}
\examples{
data(locations_testing)
graphics::par(mar = c(0, 0, 0, 0) + 0.1)
DrawEyeHeatmap(poster = system.file("extdata", "poster_colored.jpg", package = "EyeTrackR"),
locations = locations_testing)
}
\author{
Chunyang Li < lichunyang1990@hotmail.com >
}
