% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkedMicroposterPlot.R
\name{DrawEyeLMPlot}
\alias{DrawEyeLMPlot}
\title{DrawEyeLMPlot}
\usage{
DrawEyeLMPlot(poster.loc, grayscale = TRUE, AutomaticLayout = TRUE, data,
  posterVBorders, Layout = NA, MedianRow = FALSE, partitioning = 1,
  sortby = "visits", decreasing = TRUE, panel.types = c("poster",
  "legend", "dot", "dot", "boxplot"), panel.width = c(2.9, 2.9, 2.9, 2.9,
  2.9), hdColors = rev(RColorBrewer::brewer.pal(5, "Accent")),
  columns.att = list(list(col.num = 1, cumulate = TRUE), list(col.num = 2,
  header = "Areas of Interest", header.size = 0.7, point.size = 0.96, text.size
  = 0.7, text.font = 1), list(col.num = 3, panel.data = "visits.num", header =
  "Number of Visits", text.size = 0.7, point.size = 0.96, axis.ticks = NA,
  axis.labels = NA), list(col.num = 4, panel.data = "visits", header =
  "Length of Visits (sec)", text.size = 0.7, point.size = 0.96, axis.ticks = NA,
  axis.labels = NA), list(col.num = 5, panel.data = "pupildat", header =
  "Pupil Radius in Pixels",      text.size = 0.7, axis.ticks = NA, axis.labels =
  NA)), main.title = "Eye Tracking: Looking at Poster Statistics",
  title.cex = 1.08)
}
\arguments{
\item{poster.loc}{The jpg image.}

\item{grayscale}{Whether to use grayscale image for the microposter or not. Default is TRUE.}

\item{AutomaticLayout}{Whether to use the automatic layout or not. Default is TRUE.}

\item{data}{A list of summarized data by \code{\link{GetPosterData}}.}

\item{posterVBorders}{The Area of Interest border information data created by \code{\link{DrawAOIs}}.}

\item{Layout}{Self defined layout: a vector contains how many observations in each panel. Default is NA.}

\item{MedianRow}{Whether have a separate row indicate the meidian or not, if possible. Default is TRUE.}

\item{partitioning}{Default is 1. See Symanzik, J. and Carr, D.B., 2008.
Interactive linked micromap plots for the display of geographically referenced statistical data.
In Handbook of data visualization (pp. 267-294). Springer Berlin Heidelberg.}

\item{sortby}{The name of the variable to sort by. Default is "visits".
Other options are "pupildat", "speed", and "visits.num".}

\item{decreasing}{Whether sort the data by a decreasing order or not. Default is TRUE.}

\item{panel.types}{A vector of the plot types for each columns.
The possible plot types are "poster", "legend", "dot", and "boxplot".
Default is c("poster", "legend", "dot", "dot", "boxplot").}

\item{panel.width}{A vector of the width for each column. Default is c(2.9, 2.7, 3, 3, 3).}

\item{hdColors}{A vector of colors for the observations for each panel. It is recommended to
provide at least six different colors. The median row (if exists) takes the last color listed.
Default is c(rev(RColorBrewer::brewer.pal(5, "Accent")), "#000000").}

\item{columns.att}{A list of panel attributes to be specified, including: col.num, cumulate, header, header.size,
point.size, text.size, panel.data, axis.ticks, point.size, and axis.labels.}

\item{main.title}{the title of the plot. Default is "Eye Tracking: Looking at Poster Statistics".}

\item{title.cex}{the size of the title. Default is 1.08.}

\item{col.num}{The number of colomn.}

\item{cumulate}{Determines whether to hightlight the AOIs investigated from the previous panels. Default is TRUE.}

\item{header}{The title for the column.}

\item{header.size}{The text size of the column title. Default is 0.7.}

\item{point.size}{The size of the point if there is any. Default is 0.96.}

\item{text.size}{The size of the text for the axis. Default is 0.7.}

\item{text.font}{The font of the text for the axis. Default is 1.}

\item{panel.data}{The input data for the panel. e.g. "visits", "pupildat", and "visits.num".}

\item{axis.ticks}{A vector that indicate where to label the axis. If it is NA, the axis ticks are
automatly determined. Default is NA.}

\item{axis.labels}{The labels on the axis. If it is NA, the axis labels are automatly determined. Default is NA.}
}
\value{
Returns a linked microposter plot.
}
\description{
Creates a linked microposter plot.
}
\examples{
# An example using the automatic layout
data("AOIName", "posterdat.all")
library(RColorBrewer)
DrawEyeLMPlot(poster.loc = system.file("extdata", "poster_colored.jpg", package = "EyeTrackR"),
                       data = posterdat.all,
                       posterVBorders = AOIName,
                       panel.types = c("poster", "legend", "dot", "dot", "boxplot"),
                       panel.width = c(2.9, 2.7, 3, 3, 3),
                       hdColors = c(rev(RColorBrewer::brewer.pal(5, "Accent")), "#000000"),
                       columns.att = list(
                         list(col.num = 1, cumulate = TRUE),
                         list(col.num = 2, header = "Areas of Interest", header.size = 0.7,
                         point.size = 0.96, text.size = 0.7, text.font = 1),
                         list(col.num = 4, panel.data = "visits.num",
                         header = "Number of Visits", text.size = 0.7, point.size = 0.96,
                         axis.ticks = NA, axis.labels = NA),
                         list(col.num = 3, panel.data = "visits",
                         header = "Length of Visits (sec)", text.size = 0.7, point.size = 0.96,
                          axis.ticks = NA, axis.labels = NA),
                         list(col.num = 5, panel.data = "pupildat",
                         header = "Pupil Radius in Pixels", text.size = 0.7,
                          axis.ticks = NA, axis.labels = NA)
                       )
)

# Run an example using self defined layout with a median row
data("AOIName", "posterdat.all")
library(RColorBrewer)
DrawEyeLMPlot(poster.loc = system.file("extdata", "poster_colored.jpg", package = "EyeTrackR"),
                        data = posterdat.all,
                        posterVBorders = AOIName,
                        Layout = c(6, 1, 6),
                        MedianRow = TRUE,
                        panel.types = c("poster", "legend", "dot", "dot", "boxplot"),
                        panel.width = c(2.9, 2.7, 3, 3, 3),
                        hdColors = c(rev(RColorBrewer::brewer.pal(5, "Accent")), "#000000"),
                        columns.att = list(
                          list(col.num = 1, cumulate = TRUE),
                          list(col.num = 2, header = "Areas of Interest", header.size = 0.7,
                          point.size = 0.96, text.size = 0.7, text.font = 1),
                          list(col.num = 4, panel.data = "visits.num",
                          header = "Number of Visits", text.size = 0.7, point.size = 0.96,
                          axis.ticks = NA, axis.labels = NA),
                          list(col.num = 3, panel.data = "visits",
                          header = "Length of Visits (sec)", text.size = 0.7, point.size = 0.96,
                          axis.ticks = NA, axis.labels = NA),
                          list(col.num = 5, panel.data = "pupildat",
                          header = "Pupil Radius in Pixels", text.size = 0.7,
                          axis.ticks = NA, axis.labels = NA)
                        )
)


# Run an example with four statistical columns
data("AOIName", "posterdat.all")
# Remove the outliers of the speed
speed = posterdat.all[[2]]
posterdat.all[[2]] = lapply(speed, function(x) x[x < 100])
library(RColorBrewer)
DrawEyeLMPlot(poster.loc = system.file("extdata", "poster_colored.jpg", package = "EyeTrackR"),
        data = posterdat.all,
        posterVBorders = AOIName,
        MedianRow = TRUE,
        panel.types = c("poster", "legend", "dot", "dot", "boxplot", "boxplot"),
        panel.width = c(2.9, 2.7, 3, 3, 3, 3),
        hdColors = c(rev(RColorBrewer::brewer.pal(5, "Accent")), "#000000"),
        columns.att = list(
          list(col.num = 1, cumulate = TRUE),
          list(col.num = 2, header = "Areas of Interest", header.size = 0.7,
          point.size = 0.96, text.size = 0.7, text.font = 1),
          list(col.num = 4, panel.data = "visits.num", header = "Number of Visits",
          text.size = 0.7, point.size = 0.96, axis.ticks = NA, axis.labels = NA),
          list(col.num = 3, panel.data = "visits", header = "Length of Visits (sec)",
          text.size = 0.7, point.size = 0.96, axis.ticks = NA, axis.labels = NA),
          list(col.num = 5, panel.data = "pupildat", header = "Pupil Radius in Pixels",
          text.size = 0.7, axis.ticks = NA, axis.labels = NA),
          list(col.num = 6, panel.data = "speed", header = "Eye Speed in Pixels",
          text.size = 0.7, axis.ticks = NA, axis.labels = NA)
        )
)
}
\author{
Chunyang Li < lichunyang1990@hotmail.com >
}
