% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScanpathMicroposterPlot2.R
\name{DrawEyeLSMPlot}
\alias{DrawEyeLSMPlot}
\title{DrawEyeLSMPlot}
\usage{
DrawEyeLSMPlot(poster.loc, locations, grayscale = TRUE, adjust = FALSE,
  AutomaticLayout = TRUE, lambda = 10, frequency = 30, Layout = NA,
  MedianRow = FALSE, partitioning = 2, sortby = "sequence",
  decreasing = FALSE, panel.types = c("poster", "dot", "boxplot"),
  panel.width = c(2.9, 2.9, 2.9),
  hdColors = c(rev(RColorBrewer::brewer.pal(5, "Accent")), "#000000"),
  columns.att = list(list(col.num = 1, sequence.label = TRUE, label.size = 1,
  circle.size = 5, line = TRUE, header = "Scanpath Microposters"), list(col.num
  = 2, panel.data = "visits", header = "Length of Visits (sec)", text.size =
  0.7, point.size = 0.96, axis.ticks = NA, axis.labels = NA), list(col.num = 3,
  panel.data = "pupildat", header = "Pupil Radius in Pixels", text.size = 0.7,
  axis.ticks = NA, axis.labels = NA)),
  main.title = "Eye Tracking: Looking at Poster Statistics",
  title.cex = 1.08)
}
\arguments{
\item{poster.loc}{The jpg image.}

\item{locations}{The directory and name of the csv file of the eye tracking data coordinates.}

\item{grayscale}{Whether to use grayscale image for the microposter or not. Default is TRUE.}

\item{adjust}{Adjust the difference in the coordinates from the results returned by Matlab. Default is TRUE.}

\item{AutomaticLayout}{Whether to use the automatic layout or not. Default is TRUE.}

\item{lambda}{The threshold for fixation detection. The smaller lambda is, the more fixations it detects. Default is 10.}

\item{frequency}{The frequency of the recording. Default is 30.}

\item{Layout}{Self defined layout: a vector contains how many observations in each panel. Default is NA.}

\item{MedianRow}{Whether have a separate row indicate the meidian or not, if possible. Default is FALSE.}

\item{partitioning}{Default is 2. See Symanzik, J. and Carr, D.B., 2008.
Interactive linked micromap plots for the display of geographically referenced statistical data.
In Handbook of data visualization (pp. 267-294). Springer Berlin Heidelberg.}

\item{sortby}{The name of the variable to sort by. Default is "visits".
Other options are "pupildat" and "visits".}

\item{decreasing}{Whether sort the data by a decreasing order or not. Default is TRUE.}

\item{panel.types}{A vector of the plot types for each columns.
The possible plot types are "poster", "legend", "dot", and "boxplot".
Default is c("poster", "dot", "boxplot").}

\item{panel.width}{A vector of the width for each column. Default is c(2.9, 2.9, 2.9).}

\item{hdColors}{A vector of colors for the observations for each panel. It is recommended to
provide at least six different colors. The median row (if exists) takes the last color listed.
Default is c(rev(RColorBrewer::brewer.pal(5, "Accent")), "#000000").}

\item{columns.att}{A list of panel attributes to be specified, including: col.num, cumulate, header, header.size,
point.size, text.size, panel.data, axis.ticks, point.size, and axis.labels.}

\item{main.title}{the title of the plot. Default is "Eye Tracking: Looking at Poster Statistics".}

\item{title.cex}{the size of the title. Default is 1.08.}

\item{col.num}{The number of colomn.}

\item{sequence.label}{A logical variable that determines whether to label the fixation sequences in the poster or not.
Default is TRUE.}

\item{label.size}{The size of the number that indicates the fixation sequences, if the sequence.label is TRUE.
Default is 1.}

\item{circle.size}{The size of the circles that represent the fixation points. Default is 5.}

\item{line}{A logical variable indicates whether to connect the fixation points with lines or not.}

\item{header}{The title for the column.}

\item{point.size}{The size of the point if there is any. Default is 0.96.}

\item{text.size}{The size of the text for the axis. Default is 0.7.}

\item{panel.data}{The input data for the panel. e.g. "visits", "pupildat", and "visits.num".}

\item{axis.ticks}{A vector that indicate where to label the axis. If it is NA, the axis ticks are
automatly determined. Default is NA.}

\item{axis.labels}{The labels on the axis. If it is NA, the axis labels are automatly determined. Default is NA.}
}
\value{
Returns a linked scanpath microposter plot without predefined AOIs.
}
\description{
Creates a linked scanpath microposter plot without predefined AOIs.
}
\details{
The fixation points are identified by \code{\link[saccades]{detect.fixations}}.
}
\examples{
data("locations_testing")


DrawEyeLSMPlot(poster.loc = system.file("extdata", "poster_colored.jpg", package = "EyeTrackR"),
          locations = locations_testing,
          columns.att = list(
            list(col.num = 1, sequence.label = TRUE, label.size = 1, circle.size = 5,
             line = TRUE, header = "Scanpath Microposters"),
            list(col.num = 2, panel.data = "visits", header = "Length of Visits (sec)",
             text.size = 0.7, point.size = 0.96, axis.ticks = NA, axis.labels = NA),
            list(col.num = 3, panel.data = "pupildat", header = "Pupil Radius in Pixels",
             text.size = 0.7, axis.ticks = NA, axis.labels = NA)
          )
)

# Example two: change the layout to self-defined layout and enabled the median row
# set lambda to 8 in order to have more fixations identified.

data("locations_testing")


DrawEyeLSMPlot(poster.loc = system.file("extdata", "poster_colored.jpg", package = "EyeTrackR"),
          locations = locations_testing,
          lambda = 8,
          Layout = c(6, 5, 5, 1, 5, 5, 6),
          hdColors = c('#a6cee3','#1f78b4','#b2df8a','#33a02c','#fb9a99','#e31a1c',
          '#fdbf6f','#ff7f00','#cab2d6','#6a3d9a','#ffff99','#b15928', "#000000"),
          columns.att = list(
            list(col.num = 1, sequence.label = TRUE, label.size = 1, circle.size = 5,
             line = TRUE, header = "Scanpath Microposters"),
            list(col.num = 2, panel.data = "visits", header = "Length of Visits (sec)",
             text.size = 0.7, point.size = 0.96, axis.ticks = NA, axis.labels = NA),
            list(col.num = 3, panel.data = "pupildat", header = "Pupil Radius in Pixels",
             text.size = 0.7, axis.ticks = NA, axis.labels = NA)
          )
)


# Example three: Sort by length of visits and remove the connecting line between fixations

data("locations_testing")
library(RColorBrewer)

DrawEyeLSMPlot(poster.loc = system.file("extdata", "poster_colored.jpg", package = "EyeTrackR"),
          locations = locations_testing,
          sortby = "visits",
          columns.att = list(
            list(col.num = 1, sequence.label = TRUE, label.size = 1, circle.size = 5,
             line = FALSE, header = "Scanpath Microposters"),
            list(col.num = 2, panel.data = "visits", header = "Length of Visits (sec)",
             text.size = 0.7, point.size = 0.96, axis.ticks = NA, axis.labels = NA),
            list(col.num = 3, panel.data = "pupildat", header = "Pupil Radius in Pixels",
             text.size = 0.7, axis.ticks = NA, axis.labels = NA)
          )
)

}
\author{
Chunyang Li < lichunyang1990@hotmail.com >
}
