% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScanpathMicroposterPlot.R
\name{DrawEyeLTMPlot}
\alias{DrawEyeLTMPlot}
\title{DrawEyeLTMPlot}
\usage{
DrawEyeLTMPlot(poster.loc, grayscale = TRUE, AutomaticLayout = TRUE, data,
  posterVBorders, Layout = NA, MedianRow = TRUE, partitioning = 1,
  sortby = "sequence", decreasing = TRUE, colorby = "AOIs",
  panel.types = c("poster", "legend", "dot", "boxplot", "boxplot"),
  panel.width = c(2.9, 2.9, 2.9, 2.9, 2.9),
  hdColors = c(rev(RColorBrewer::brewer.pal(5, "Accent")), "#000000"),
  columns.att = list(list(col.num = 1, sequence.label = TRUE, label.size = 1),
  list(col.num = 2, header = "Areas of Interest", header.size = 0.7, point.size
  = 0.96, text.size = 0.7, text.font = 1), list(col.num = 3, panel.data =
  "visits", header = "Length of Visits (sec)", text.size = 0.7, point.size =
  0.96, axis.ticks = NA, axis.labels = NA), list(col.num = 5, panel.data =
  "speed", header = "Eye Movement Speed in Pixels", text.size = 0.7, axis.ticks
  = NA, axis.labels = NA), list(col.num = 4, panel.data = "pupildat",     
  header = "Pupil Radius in Pixels", text.size = 0.7, axis.ticks = NA,
  axis.labels = NA)), main.title = "Linked Timeline Microposter Plot",
  title.cex = 1.08)
}
\arguments{
\item{poster.loc}{The jpg image.}

\item{grayscale}{Whether to use grayscale image for the microposter or not. Default is TRUE.}

\item{AutomaticLayout}{Whether to use the automatic layout or not. Default is TRUE.}

\item{data}{A list of summarized data by \code{\link{GetPosterData}}.}

\item{posterVBorders}{The Area of Interest border information data created by \code{\link{DrawAOIs}}.}

\item{Layout}{Self defined layout: a vector contains how many observations in each panel. Default is NA.}

\item{MedianRow}{Whether have a separate row indicate the meidian or not, if possible. Default is TRUE.}

\item{partitioning}{Default is 1. See Symanzik, J. and Carr, D.B., 2008.
Interactive linked micromap plots for the display of geographically referenced statistical data.
In Handbook of data visualization (pp. 267-294). Springer Berlin Heidelberg.}

\item{sortby}{The name of the variable to sort by. Default is "sequence".
Other options are "pupildat", "speed", and "visits".}

\item{decreasing}{Whether sort the data by a decreasing order or not. Default is TRUE.}

\item{colorby}{Two options: "AOIs" or "panels". Default is "AOIs".
Determines either to color the AOIs by AOIs (every AOI has
 a different color if there are plenty of colors provided) or by panels
 (the color is arranged by different panels and get recycled for each panel).}

\item{panel.types}{A vector of the plot types for each columns.
The possible plot types are "poster", "legend", "dot", and "boxplot".
Default is c("poster", "legend", "dot", "dot", "boxplot").}

\item{panel.width}{A vector of the width for each column. Default is c(2.9, 2.7, 3, 3, 3).}

\item{hdColors}{A vector of colors for the observations for each panel. It is recommended to
provide at least six different colors. #' Default is c(rev(RColorBrewer::brewer.pal(5, "Accent")), "#000000").}

\item{columns.att}{A list of panel attributes to be specified, including: col.num, header, header.size,
point.size, text.size, panel.data, axis.ticks, point.size, and axis.labels.}

\item{main.title}{the title of the plot. Default is "Linked Timeline Microposter Plot".}

\item{title.cex}{the size of the title. Default is 1.08.}

\item{col.num}{The number of colomn.}

\item{header}{The title for the column.}

\item{header.size}{The text size of the column title. Default is 0.7.}

\item{point.size}{The size of the point if there is any. Default is 0.96.}

\item{text.size}{The size of the text for the axis. Default is 0.7.}

\item{text.font}{The font of the text for the axis. Default is 1.}

\item{panel.data}{The input data for the panel. e.g. "visits", "pupildat", and "visits.num".}

\item{axis.ticks}{A vector that indicate where to label the axis. If it is NA, the axis ticks are
automatly determined. Default is NA.}

\item{axis.labels}{The labels on the axis. If it is NA, the axis labels are automatly determined. Default is NA.}
}
\value{
Returns a linked scanpath microposter plot.
}
\description{
Creates a linked scanpath microposter plot.
}
\examples{
### Example one
data("AOIName", "scanpathdat.all")
# Remove the outliers of the speed
speed = scanpathdat.all[[2]]
scanpathdat.all[[2]] = lapply(speed, function(x) x[x < 100])
DrawEyeLTMPlot(poster.loc = system.file("extdata", "poster_colored.jpg", package = "EyeTrackR"),
         data = scanpathdat.all,
         posterVBorders = AOIName,
         partitioning = 2,
         decreasing = FALSE,
         panel.types = c("poster", "legend", "dot", "boxplot", "boxplot"),
         panel.width = c(2, 2, 2.5, 2.5, 2.5),
         hdColors = c('#a6cee3','#1f78b4','#b2df8a','#33a02c','#fb9a99',
         '#e31a1c','#fdbf6f','#ff7f00','#cab2d6','#6a3d9a','#ffff99','#b15928'),
         columns.att = list(
         list(col.num = 1, sequence.label = TRUE, label.size = 1),
         list(col.num = 2, header = "Areas of Interest", header.size = 0.7,
         point.size = 0.96, text.size = 0.7, text.font = 1),
         list(col.num = 3, panel.data = "visits", header = "Length of Visits (sec)",
         text.size = 0.7, point.size = 0.96, axis.ticks = NA, axis.labels = NA),
         list(col.num = 4, panel.data = "pupildat", header = "Pupil Radius in Pixels",
         text.size = 0.7, axis.ticks = NA, axis.labels = NA),
         list(col.num = 5, panel.data = "speed", header = "Eye Movement Speed in Pixels",
         text.size = 0.7, axis.ticks = NA, axis.labels = NA)
                            ),
         main.title = "Linked Timeline Microposter Plot"
)


# Example two: only two statistical panels; arrange panels by color;
# rearrange the panels

data("AOIName", "scanpathdat.all")
# Remove the outliers of the speed
speed = scanpathdat.all[[2]]
scanpathdat.all[[2]] = lapply(speed, function(x) x[x < 100])

library(RColorBrewer)

DrawEyeLTMPlot(poster.loc = system.file("extdata", "poster_colored.jpg", package = "EyeTrackR"),
         data = scanpathdat.all,
         posterVBorders = AOIName,
         partitioning = 2,
         decreasing = FALSE,
         panel.types = c("poster", "legend", "dot", "boxplot"),
         panel.width = c(2.5, 2.5, 2, 2),
         hdColors = rev(RColorBrewer::brewer.pal(5, "Accent")),
         colorby = "panels",
         columns.att = list(
           list(col.num = 4, sequence.label = TRUE, label.size = 1),
           list(col.num = 3, header = "Areas of Interest", header.size = 0.7,
           point.size = 0.96, text.size = 0.7, text.font = 1),
           list(col.num = 2, panel.data = "visits", header = "Length of Visits (sec)",
           text.size = 0.7, point.size = 0.96, axis.ticks = NA, axis.labels = NA),
           list(col.num = 1, panel.data = "pupildat", header = "Pupil Radius in Pixels",
           text.size = 0.7, axis.ticks = NA, axis.labels = NA)
         )
)



# Example 3: rearrange panels and sort by pupil radius

data("AOIName", "scanpathdat.all")
# Remove the outliers of the speed
speed = scanpathdat.all[[2]]
scanpathdat.all[[2]] = lapply(speed, function(x) x[x < 100])

library(RColorBrewer)

DrawEyeLTMPlot(poster.loc = system.file("extdata", "poster_colored.jpg", package = "EyeTrackR"),
         data = scanpathdat.all,
         posterVBorders = AOIName,
         Layout = c(5, 3, 3, 5),
         partitioning = 2,
         sortby = "pupildat",
         decreasing = TRUE,
         panel.types = c("poster", "legend", "dot", "boxplot", "boxplot"),
         panel.width = c(2, 2, 2.5, 2.5, 2.5),
         hdColors = c('#a6cee3','#1f78b4','#b2df8a','#33a02c','#fb9a99','#e31a1c',
         '#fdbf6f','#ff7f00','#cab2d6','#6a3d9a','#ffff99','#b15928'),
         columns.att = list(
           list(col.num = 1, sequence.label = TRUE, label.size = 1),
           list(col.num = 2, header = "Areas of Interest", header.size = 0.7,
           point.size = 0.96, text.size = 0.7, text.font = 1),
           list(col.num = 3, panel.data = "visits", header = "Length of Visits (sec)",
            text.size = 0.7, point.size = 0.96, axis.ticks = NA, axis.labels = NA),
           list(col.num = 5, panel.data = "speed", header = "Eye Movement Speed in Pixels",
            text.size = 0.7, axis.ticks = NA, axis.labels = NA),
           list(col.num = 4, panel.data = "pupildat", header = "Pupil Radius in Pixels",
           text.size = 0.7, axis.ticks = NA, axis.labels = NA)
         )
)

}
\author{
Chunyang Li < lichunyang1990@hotmail.com >
}
