% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractCoordinates.R
\name{ExtractCoordinates}
\alias{ExtractCoordinates}
\title{ExtractCoordinates}
\usage{
ExtractCoordinates(framedir, poster, coordinate = NA, outputcsv = NA,
  outputimg = NA, ibegin, iend)
}
\arguments{
\item{framedir}{the directory of the video frames.}

\item{poster}{the clear version of the poster image.}

\item{coordinate}{the x and y coordinates of the crosshair in terms of the
scene (Only two columns and no headers). NA if the coordinates are not available.
The default is NA.}

\item{outputcsv}{the directory and name of the output csv file. The default is the
current working directory.}

\item{outputimg}{the directory of the output image file, put NaN if you don't
want the output matching images saved. The default is the current working directory.}

\item{ibegin}{the starting frame.}

\item{iend}{the ending frame.}
}
\value{
Returns the coordinates of the crosshair in terms of the poster and
a folder of the matched poster images.
}
\description{
Extract the coordinates of the crosshair in terms of the poster from the video frames
}
\examples{
\dontrun{
# Put the right version of Matlab that installed on your computer
library(matlabr)
options(matlab.path = "/Applications/MATLAB_R2016b.app/bin")
Check having Matlab or not
have_matlab()

ExtractCoordinates(framedir = system.file("extdata", "Frames", package = "EyeTrackR"),
poster = system.file("extdata", "poster_colored.jpg", package = "EyeTrackR"),
coordinate = NA,
ibegin = 1,
iend = 5)
file.exists("locations.csv")
}
\dontrun{
# Put the right version of Matlab that installed on your computer
options(matlab.path = "/Applications/MATLAB_R2016b.app/bin")
library(matlabr)
# Check having Matlab or not
have_matlab()

ExtractCoordinates(framedir = system.file("extdata", "Frames", package = "EyeTrackR"),
poster = system.file("extdata", "poster_colored.jpg", package = "EyeTrackR"),
coordinate = crosshair,
ibegin = 1,
iend = 5)
file.exists("locations.csv")
}
}
