\name{eyeTrackR-package}
\alias{eyeTrackR-package}
\alias{eyeTrackR}
\docType{package}
\title{
EyeTrackR: Functions for organising and analysing eye-tracking data
}
\description{
A variety of different functions for handling output from eyetrackers and providing means for analyses on a wide range of standard measures.
}
\details{
\tabular{ll}{
Package: \tab eyeTrackR\cr
Type: \tab Package\cr
Version: \tab 0.01\cr
Date: \tab 2012-03-25\cr
License: \tab GNU GPL. \cr Also institutional disclaimer: These free software programs provided by The University of Southampton may be freely distributed, provided that no charge above the cost of distribution is levied, and that the disclaimer below is always attached to it. The programs are provided 'as is' without any guarantees or warranty. Although the author has attempted to find and correct any bugs in the free software programs, the author is not responsible for any damage or losses of any kind caused by the use or misuse of the programs. The author is under no obligation to provide support, service, corrections, or upgrades to the free software programs.\cr
}

}
\author{
Hayward J. Godwin

Maintainer: Hayward J. Godwin <hg102@soton.ac.uk>
}
