% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSummary.R
\name{GetAOITimelineData}
\alias{GetAOITimelineData}
\title{GetAOITimelineData}
\usage{
GetAOITimelineData(posterVBorders = posterVBorders, locations = locations,
  threshold = 3, frequency = 30, adjust = TRUE)
}
\arguments{
\item{posterVBorders}{The output csv file from the \code{\link{DrawAOIs}} function.}

\item{locations}{The location and name of the csv file of the eye tracking data coordinates.}

\item{threshold}{Defines for at least how many images the coordinates fall into a certain AOI to be considered as a visit.}

\item{frequency}{The frequency of the eye tracker.}

\item{adjust}{Adjust the difference in the coordinates from the results returned by Matlab. Default is TRUE.}
}
\value{
A list of three or four data frames and lists with summarized statistics of scanpath data according to the AOIs drawn.
}
\description{
Summarize scanpath data according to the AOIs drawn
}
\details{
The fist element of the list is a data frame contains the length of visits for each visit and the temporal sequence of the visits;
The second element of the list is a list of eye movement speed in pixels for the visit at each AOIs;
The third element of the list is the summary of the length and times the participant is looking at the areas not included in any AOIs;
The fourth element of the list is a list of pupil radiuses for each visit at the coorespoinding AOI, if the pupil radius is provided in the location csv file.
}
\examples{
scanpathdat.all = GetAOITimelineData(posterVBorders = AOIName,
locations = locations_testing)
}
\author{
Chunyang Li < lichunyang1990@hotmail.com >
}
