% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{locations_testing}
\alias{locations_testing}
\title{locations_testing}
\format{A data frame with 1647 rows and 3 variables.


\describe{
\item{V1}{x coordinates}
\item{V2}{y coordinates}
\item{V3}{pupil radius in pixels}
}}
\usage{
data(locations_testing)
}
\description{
The coordinates returned by Matlab from the data collectd from looking at the example poster:
the participant looked at the logo, title and author, image top left and image bottom middle
for two seconds and the rest of the area of interests for six seconds.
The viewing sequence is: logo, title and author, image top right, data set information, accuracy tables,
conclusion, introduction, image bottomn middle, confusion matrices, variable importance, classification,
and acknowledgements.
}
\keyword{datasets}
