% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORGANISE__functions.R
\name{organise.contingencies.descriptives}
\alias{organise.contingencies.descriptives}
\title{Descriptive statistics of fixation contingencies.}
\usage{
organise.contingencies.descriptives(fixreport_df)
}
\arguments{
\item{fixreport_df}{Fixation report.}
}
\value{
Output to console.
}
\description{
Descriptive statistics of fixation contingencies.
}
\examples{
data(fixationreport)
data(messagereport)

# REPLACE SPACES IN MESSAGES
messagereport <- organise.message.replace_spaces(messagereport)

# TAKE A LOOK AT THE MESSAGES WE HAVE
organise.message.descriptives(messagereport)

# MARKUP - SYNCTIME IS THE START OF THE TRIAL AND LEADIN IS A FIXATION CROSS BEFOREHAND
# THERE IS A TRIAL END BVUT WE DON'T WORRY ABOUT THAT SINCE WE USE THE RESPONSE MESSAGE INSTEAD
foo <- organise.message.markup(message_df=messagereport, fixreport_df = fixationreport, 
                               message="LEADIN")
foo <- organise.message.markup(message_df=messagereport, fixreport_df = foo, 
                               message="SYNCTIME")

# NOW DO ACCURACY AND RT MARKUP
foo <- organise.responses.markup(foo, "CORRECT_ANSWER")

# NOW MARK UP FIXATION CONTINGENCIES
foo<-organise.message.fix_contingencies(foo, list("LEADIN", "SYNCTIME", "RESPONSE_TIME"))

# RANDOM TRIAL TO CHECK THINGS OUT
organise.checks.random_trial(foo)

# FIX CONTINGENCIES
organise.contingencies.descriptives(foo)
}
