\name{organise.fixreport.remove_rare}
\alias{organise.fixreport.remove_rare}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Fixation report - remove rare columns.
}
\description{
Removes rarely-used columns in the fixation report.
}
\usage{
organise.fixreport.remove_rare(fixreport_df)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{fixreport_df}{
Input fixation report.
}
}
\details{
Useful if you have a large dataframe and are low on RAM.
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Hayward Godwin
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
##---- TBA
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
