\name{organise.message.first_fixation_time_period_info}
\alias{organise.message.first_fixation_time_period_info}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
First fixation time period
}
\description{
Adds information on the first fixation in a time period to a fixation report.
}
\usage{
organise.message.first_fixation_time_period_info(fixreport_df, time_period)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{fixreport_df}{
Fixation report.
}
  \item{time_period}{
Time period name.
}
}
\details{
You will need to set up the time periods first before using this.
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Hayward Godwin
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
##---- TBA
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
