\name{organise.message.return_specific}
\alias{organise.message.return_specific}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Return a specific message event.
}
\description{
Returns instances of a specific message event occurring. Used for checking messages.
}
\usage{
organise.message.return_specific(message_df, fixreport_df, message, show_working = FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{message_df}{
Message report dataframe.
}
  \item{fixreport_df}{
Fixation report dataframe.
}
  \item{message}{
Message text (e.g., "SYNCTIME")
}
  \item{show_working}{
TRUE or FALSE for whether to show internal calculations or not. Defaults to FALSE.
}
}
\details{
Helper function.
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Hayward Godwin
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
##---- TBA
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
