% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORGANISE__functions.R
\name{organise.responses.markup}
\alias{organise.responses.markup}
\title{Mark up responses into a fixation report.}
\usage{
organise.responses.markup(fixreport_df, correct_answer_column)
}
\arguments{
\item{fixreport_df}{Fixation report}

\item{correct_answer_column}{The column in the fixation report containing the correct button response number (1-7).}
}
\value{
Updated fixation report as a data.table.
}
\description{
Mark up responses into a fixation report.
}
\examples{

data(fixationreport)
data(messagereport)

# REPLACE SPACES IN MESSAGES
messagereport <- organise.message.replace_spaces(messagereport)

# TAKE A LOOK AT THE MESSAGES WE HAVE
organise.message.descriptives(messagereport)

# MARKUP - SYNCTIME IS THE START OF THE TRIAL AND LEADIN IS A FIXATION CROSS BEFOREHAND
# THERE IS A TRIAL END BVUT WE DON'T WORRY ABOUT THAT SINCE WE USE THE RESPONSE MESSAGE INSTEAD
foo <- organise.message.markup(message_df=messagereport, fixreport_df = fixationreport, 
   message="LEADIN")
foo <- organise.message.markup(message_df=messagereport, fixreport_df = foo, 
   message="SYNCTIME")

# NOW DO ACCURACY AND RT MARKUP
foo <- organise.responses.markup(foo, "CORRECT_ANSWER")
}
