% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{posterdat.all}
\alias{posterdat.all}
\title{posterdat.all}
\format{A list with three lists.
\describe{
\item{[1]}{a dataframe with three columns:}
\describe{
\item{section}{AOI names.}
\item{visits}{time the participant spent on each AOI (in seconds).}
\item{visits.num}{how many times the participant has visited each AOI.}
}
\item{[2]}{a list of eye movement speed for each AOIs (each AOI is a list).}
\item{[3]}{A list of pupil radius for each AOIs (each AOI is a list).}
}}
\usage{
data(posterdat.all)
}
\description{
A list of summarized eye movement statics with function \code{get.posterdat}.
}
\keyword{datasets}
