% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertImages_DrawAOIs.R
\name{ResizeImg}
\alias{ResizeImg}
\title{ResizeImg}
\usage{
ResizeImg(poster, size_x = 640, size_y = 480, resized_poster)
}
\arguments{
\item{poster}{the location and name of the poster image file.}

\item{size_x}{the width of the new image. Default is 640.}

\item{size_y}{the height of the new image. Default is 480.}

\item{resized_poster}{the directory and the name of the output jpg file.}
}
\value{
A resized the image in the specified dimensions
}
\description{
Resize the image into the specified dimensions
}
\examples{
ResizeImg(poster = system.file("extdata", "poster_colored.jpg", package = "EyeTrackR"),
size_x = 500, size_y = 400, resized_poster = "poster_colored2.jpg")
}
\author{
Chunyang Li < lichunyang1990@hotmail.com >
}
