% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{scanpathdat.all}
\alias{scanpathdat.all}
\title{scanpathdat.all}
\format{A list with three lists.
\describe{
\item{[1]}{a dataframe with three columns:}
\describe{
\item{section}{AOI names.}
\item{visits}{time the participant spent on each AOI (in seconds).}
\item{visits.num}{how many times the participant has visited each AOI.}
}
\item{[2]}{a list of eye movement speed for each visit (each visit is a list).}
\item{[3]}{a list of list with data of people looking at the areas that are not defined
as an AOI:}
\describe{
\item{[1]}{a list with two elements: time spent in the areas and the number of visits for the areas.}
\item{[2]}{a list of eye movement speed in the areas.}
\item{[3]}{a list of pupil radius in the areas.}
}
\item{[4]}{a list of pupil radius for each visit (each visit is a list).}
}}
\usage{
data(scanpathdat.all)
}
\description{
A list of summarized eye movement statics with function \code{get.scanpathdat}.
}
\keyword{datasets}
