% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORGANISE__functions.R
\name{organise.fixreport.remove_rare}
\alias{organise.fixreport.remove_rare}
\title{Remove often unused columns in fixation reports.}
\usage{
organise.fixreport.remove_rare(fixreport_df)
}
\arguments{
\item{fixreport_df}{Input fixation report.}
}
\value{
Output fixation report with fewer columns.
}
\description{
Remove often unused columns in fixation reports.
}
\examples{
data(fixationreport)
fixationreport <- organise.fixreport.remove_rare(fixationreport)
}
